+++ src/services.py
@@ -28,5 +28,3 @@ def process_bank_search(search=None) -> json:
         return result
     else:
         return "Транзакции не обнаружены"
-
-
